Tweaked the CharSheet constructor. Worked on layouts.
Updated the CharSheet constructor such that it is not explicitly coupled
with the Android SDK.

Started piecing together a more finalized structure for the layouts.
Main view will be character list. Each item in the list will represent a
single character sheet.

In portrait orientation, a character sheet list item will display Name
and Soul Level (Starting Class as well, if there is room). The Add New
Character Sheet button will be at the bottom of the screen, for
convenience, as well as in the menu.

In landscape orientation Name, Soul Level, and Attribute Scores. The Add
New Character Sheet button will only be in the menu as this is intended
more as a quick comparison between characters.